# cmd/docs/swagger.yaml
definitions:
  PlantSite_internal_api_album-api_mapper.AddPlantToAlbumRequest:
    properties:
      plant_id:
        type: string
    required:
    - plant_id
    type: object
  PlantSite_internal_api_album-api_mapper.CreateAlbumRequest:
    properties:
      description:
        type: string
      name:
        type: string
      plant_ids:
        items:
          type: string
        type: array
    required:
    - description
    - name
    - plant_ids
    type: object
  PlantSite_internal_api_album-api_mapper.RemovePlantFromAlbumRequest:
    properties:
      plant_id:
        type: string
    required:
    - plant_id
    type: object
  PlantSite_internal_api_album-api_mapper.UpdateAlbumDescriptionRequest:
    properties:
      description:
        type: string
    required:
    - description
    type: object
  PlantSite_internal_api_album-api_mapper.UpdateAlbumNameRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  PlantSite_internal_api_album-api_response.GetAlbumResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      plantIDs:
        items:
          type: string
        type: array
      updated_at:
        type: string
    required:
    - created_at
    - description
    - id
    - name
    - updated_at
    type: object
  PlantSite_internal_api_album-api_response.ListAlbum:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      plantIDs:
        items:
          type: string
        type: array
      updated_at:
        type: string
    required:
    - created_at
    - description
    - id
    - name
    - updated_at
    type: object
...
  internal_api_auth-api.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_api_auth-api.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /album/add/{id}:
    post:
      consumes:
      - application/json
      description: Adds a plant to an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      - description: Add plant to album request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PlantSite_internal_api_album-api_mapper.AddPlantToAlbumRequest'
      responses:
        "200":
          description: Plant added to album successfully
        "400":
          description: Bad Request - Invalid input or missing required fields
        "401":
          description: Unauthorized - Not authorized to add plant to album
        "500":
          description: Internal Server Error - Failed to add plant to album
      summary: Add plant to album
      tags:
      - album
  /album/create:
    post:
      consumes:
      - application/json
      description: Creates a new album with the provided name and description
      parameters:
      - description: Create album request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PlantSite_internal_api_album-api_mapper.CreateAlbumRequest'
      responses:
        "200":
          description: Album created successfully
        "400":
          description: Bad Request - Invalid input or missing required fields
        "401":
          description: Unauthorized - Not authorized to create album
        "500":
          description: Internal Server Error - Failed to create album
      summary: Create an album
      tags:
      - album
  /album/delete/{id}:
    delete:
      description: Deletes an album by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Album deleted successfully
        "400":
          description: Bad Request - Invalid input
        "401":
          description: Unauthorized - Not authorized to delete album
        "500":
          description: Internal Server Error - Failed to delete album
      summary: Delete album
      tags:
      - album
  /album/description/{id}:
    put:
      consumes:
      - application/json
      description: Updates the description of an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      - description: Update album description request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PlantSite_internal_api_album-api_mapper.UpdateAlbumDescriptionRequest'
      responses:
        "200":
          description: Album description updated successfully
        "400":
          description: Bad Request - Invalid input or missing required fields
        "401":
          description: Unauthorized - Not authorized to update album description
        "500":
          description: Internal Server Error - Failed to update album description
      summary: Update album description
      tags:
      - album
  /album/get/{id}:
    get:
      description: Gets an album by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Album fetch successfully
          schema:
            $ref: '#/definitions/PlantSite_internal_api_album-api_response.GetAlbumResponse'
        "400":
          description: Bad Request - Invalid input
        "401":
          description: Unauthorized - Not authorized to get album
        "500":
          description: Internal Server Error - Failed to get album
      summary: Get album
      tags:
      - album
  /album/list:
    get:
      description: Lists all albums of authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Albums fetch successfully
          schema:
            items:
              $ref: '#/definitions/PlantSite_internal_api_album-api_response.ListAlbum'
            type: array
        "401":
          description: Unauthorized - Not authorized to list albums
        "500":
          description: Internal Server Error - Failed to list albums
      summary: List albums
      tags:
      - album
  /album/name/{id}:
    put:
      consumes:
      - application/json
      description: Updates the name of an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      - description: Update album name request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PlantSite_internal_api_album-api_mapper.UpdateAlbumNameRequest'
      responses:
        "200":
          description: Album name updated successfully
        "400":
          description: Bad Request - Invalid input or missing required fields
        "401":
          description: Unauthorized - Not authorized to update album name
        "500":
          description: Internal Server Error - Failed to update album name
      summary: Update album name
      tags:
      - album
  /album/remove/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a plant from an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      - description: Remove plant from album request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PlantSite_internal_api_album-api_mapper.RemovePlantFromAlbumRequest'
      responses:
        "200":
          description: Plant removed from album successfully
        "400":
          description: Bad Request - Invalid input or missing required fields
        "401":
          description: Unauthorized - Not authorized to remove plant from album
        "500":
          description: Internal Server Error - Failed to remove plant from album
      summary: Remove plant from album
      tags:
      - album
  /auth/login:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Authenticates a user and creates a session
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_auth-api.LoginRequest'
      responses:
        "200":
          description: Session for user created
        "400":
          description: Wrong input parameters
        "401":
          description: Auth error
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      description: Logouts a user
      responses:
        "200":
          description: User logged out
        "401":
          description: Logout error
      summary: Logouts a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Registers a new user. Requires login afterwards
      parameters:
      - description: Register credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_auth-api.RegisterRequest'
      responses:
        "200":
          description: User registered
        "400":
          description: Wrong input parameters
        "401":
          description: Auth error
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"